Continue graph decomposition by extracting entity facets surface (#296)
