--- internal/graph/entities/entities.go
@@ -1,11 +1,0 @@
-package entities
-
-import graph "github.com/writer/cerebro/internal/graph"
-
-type (
-	EntityFacetFieldDefinition = graph.EntityFacetFieldDefinition
-	EntityFacetDefinition      = graph.EntityFacetDefinition
-	EntityFacetRecord          = graph.EntityFacetRecord
-	EntityPostureClaimRecord   = graph.EntityPostureClaimRecord
-	EntityPostureSummary       = graph.EntityPostureSummary
-)
