+++ internal/graph/entities/facets.go
@@ -2,6 +2,14 @@ package entities
 
 import graph "github.com/writer/cerebro/internal/graph"
 
+type (
+	EntityFacetFieldDefinition = graph.EntityFacetFieldDefinition
+	EntityFacetDefinition      = graph.EntityFacetDefinition
+	EntityFacetRecord          = graph.EntityFacetRecord
+	EntityPostureClaimRecord   = graph.EntityPostureClaimRecord
+	EntityPostureSummary       = graph.EntityPostureSummary
+)
+
 func GetEntityFacetDefinition(id string) (EntityFacetDefinition, bool) {
 	return graph.GetEntityFacetDefinition(id)
 }
