+++ internal/graph/entities/query_test.go
@@ -7,6 +7,28 @@ import (
 	graph "github.com/writer/cerebro/internal/graph"
 )
 
+func TestGetEntityRecordDelegatesToGraphEntityReadModel(t *testing.T) {
+	g := graph.New()
+	now := time.Now().UTC()
+	g.AddNode(&graph.Node{
+		ID:        "service:payments",
+		Kind:      graph.NodeKindService,
+		Name:      "payments",
+		CreatedAt: now,
+	})
+
+	record, ok := GetEntityRecord(g, "service:payments", now, now)
+	if !ok {
+		t.Fatal("expected entity record")
+	}
+	if record.ID != "service:payments" {
+		t.Fatalf("unexpected record id %q", record.ID)
+	}
+	if record.Kind != graph.NodeKindService {
+		t.Fatalf("unexpected record kind %q", record.Kind)
+	}
+}
+
 func TestQueryEntitiesFiltersAndKnowledgeSupport(t *testing.T) {
 	g := graph.New()
 	baseAt := time.Date(2026, 3, 10, 9, 0, 0, 0, time.UTC)
